Entity Framework migration - Payments

+++ DataLayer/Payment.cs
@@ -25,7 +25,7 @@ namespace DataLayer
         public string Description { get; set; }
         public byte Day { get; set; }
         public bool Enabled { get; set; }
-        public short Months { get; set; }
+        public int Months { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<CalendarItem> CalendarItems { get; set; }
